requirements.txt
azure-storage-blob<13
duckdb<2
pandas<3
pillow<12
pyodbc<6